Function eosio_cdt_sys::check_transaction_authorization

pub unsafe extern "C" fn check_transaction_authorization(
    trx_data: *const c_char, 
    trx_size: u32, 
    pubkeys_data: *const c_char, 
    pubkeys_size: u32, 
    perms_data: *const c_char, 
    perms_size: u32
) -> i32

Checks if a transaction is authorized by a provided set of keys and permissions

@param trx_data - pointer to the start of the serialized transaction @param trx_size - size (in bytes) of the serialized transaction @param pubkeys_data - pointer to the start of the serialized vector of provided public keys @param pubkeys_size - size (in bytes) of serialized vector of provided public keys (can be 0 if no public keys are to be provided) @param perms_data - pointer to the start of the serialized vector of provided permissions (empty permission name acts as wildcard) @param perms_size - size (in bytes) of the serialized vector of provided permissions

@return 1 if the transaction is authorized, 0 otherwise