Function eosio_cdt_sys::db_idx64_find_primary

pub unsafe extern "C" fn db_idx64_find_primary(
    code: capi_name, 
    scope: u64, 
    table: capi_name, 
    secondary: *mut u64, 
    primary: u64
) -> i32

Find a table row in a secondary 64-bit integer index table by primary key

@brief Find a table row in a secondary 64-bit integer index table by primary key @param code - The name of the owner of the table @param scope - The scope where the table resides @param table - The table name @param secondary - Pointer to a uint64_t variable which will have its value set to the secondary key of the found table row @param primary - The primary key of the table row to look up @post If and only if the table row is found, *secondary will be replaced with the secondary key of the found table row @return iterator to the table row with a primary key equal to id or the end iterator of the table if the table row could not be found