Function eosio_cdt_sys::db_store_i64

pub unsafe extern "C" fn db_store_i64(
    scope: u64, 
    table: capi_name, 
    payer: capi_name, 
    id: u64, 
    data: *const c_void, 
    len: u32
) -> i32

Store a record in a primary 64-bit integer index table

@brief Store a record in a primary 64-bit integer index table @param scope - The scope where the table resides (implied to be within the code of the current receiver) @param table - The table name @param payer - The account that pays for the storage costs @param id - ID of the entry @param data - Record to store @param len - Size of data @pre data is a valid pointer to a range of memory at least len bytes long @pre *((uint64_t*)data) stores the primary key @return iterator to the newly created table row @post a new entry is created in the table