Function eosio_cdt_sys::db_update_i64

pub unsafe extern "C" fn db_update_i64(
    iterator: i32, 
    payer: capi_name, 
    data: *const c_void, 
    len: u32
)

Update a record in a primary 64-bit integer index table

@brief Update a record in a primary 64-bit integer index table @param iterator - Iterator to the table row containing the record to update @param payer - The account that pays for the storage costs (use 0 to continue using current payer) @param data - New updated record @param len - Size of data @pre data is a valid pointer to a range of memory at least len bytes long @pre *((uint64_t*)data) stores the primary key @pre iterator points to an existing table row in the table @post the record contained in the table row pointed to by iterator is replaced with the new updated record