Function eosio_cdt_sys::get_active_producers

pub unsafe extern "C" fn get_active_producers(
    producers: *mut capi_name, 
    datalen: u32
) -> u32

Gets the set of active producers.

@param producers - Pointer to a buffer of account names @param datalen - Byte length of buffer, when passed 0 will return the size required to store full output.

@return uint32_t - Number of bytes actually populated @pre producers is a pointer to a range of memory at least datalen bytes long @post the passed in producers pointer gets the array of active producers.

Example:

@code capi_name producers[21]; uint32_t bytes_populated = get_active_producers(producers, sizeof(capi_name)*21); @endcode