Function eosio_cdt_sys::get_resource_limits

pub unsafe extern "C" fn get_resource_limits(
    account: capi_name, 
    ram_bytes: *mut i64, 
    net_weight: *mut i64, 
    cpu_weight: *mut i64
)

Get the resource limits of an account

@param account - name of the account whose resource limit to get @param ram_bytes - pointer to int64_t to hold retrieved ram limit in absolute bytes @param net_weight - pointer to int64_t to hold net limit @param cpu_weight - pointer to int64_t to hold cpu limit