Crate eosio_cdt_sys

Structs

CStr
Representation of a borrowed C string.

CString
A type representing an owned, C-compatible, nul-terminated string with no nul bytes in the middle.

FromBytesWithNulError
An error indicating that a nul byte was not in the expected position.

IntoStringError
An error indicating invalid UTF-8 when converting a CString into a String.

NulError
An error indicating that an interior nul byte was found.

OsStr
Borrowed reference to an OS string (see OsString).

OsString
A type that can represent owned, mutable platform-native strings, but is cheaply inter-convertible with Rust strings.

VaList	Experimental
A wrapper for a va_list

capi_checksum160

capi_checksum256

capi_checksum512

capi_public_key

capi_signature

Enums

c_void
Equivalent to C's void type when used as a pointer.

c_void

Functions

action_data_size⚠

assert_recover_key⚠

assert_ripemd160⚠

assert_sha1⚠

assert_sha256⚠

assert_sha512⚠

cancel_deferred⚠

check_permission_authorization⚠

check_transaction_authorization⚠

current_receiver⚠

current_time⚠

db_end_i64⚠

db_find_i64⚠

db_get_i64⚠

db_idx64_store⚠

db_idx64_update⚠

db_idx64_remove⚠

db_idx64_next⚠

db_idx64_previous⚠

db_idx64_find_primary⚠

db_idx64_find_secondary⚠

db_idx64_lowerbound⚠

db_idx64_upperbound⚠

db_idx64_end⚠

db_idx128_store⚠

db_idx128_update⚠

db_idx128_remove⚠

db_idx128_next⚠

db_idx128_previous⚠

db_idx128_find_primary⚠

db_idx128_find_secondary⚠

db_idx128_lowerbound⚠

db_idx128_upperbound⚠

db_idx128_end⚠

db_idx256_store⚠

db_idx256_update⚠

db_idx256_remove⚠

db_idx256_next⚠

db_idx256_previous⚠

db_idx256_find_primary⚠

db_idx256_find_secondary⚠

db_idx256_lowerbound⚠

db_idx256_upperbound⚠

db_idx256_end⚠

db_idx_double_end⚠

db_idx_double_find_primary⚠

db_idx_double_find_secondary⚠

db_idx_double_lowerbound⚠

db_idx_double_next⚠

db_idx_double_previous⚠

db_idx_double_remove⚠

db_idx_double_store⚠

db_idx_double_update⚠

db_idx_double_upperbound⚠

db_idx_long_double_end⚠

db_idx_long_double_find_primary⚠

db_idx_long_double_find_secondary⚠

db_idx_long_double_lowerbound⚠

db_idx_long_double_next⚠

db_idx_long_double_previous⚠

db_idx_long_double_remove⚠

db_idx_long_double_store⚠

db_idx_long_double_update⚠

db_idx_long_double_upperbound⚠

db_lowerbound_i64⚠

db_next_i64⚠

db_previous_i64⚠

db_remove_i64⚠

db_store_i64⚠

db_update_i64⚠

db_upperbound_i64⚠

eosio_assert⚠

eosio_assert_code⚠

eosio_assert_message⚠

eosio_exit⚠

expiration⚠

get_account_creation_time⚠

get_action⚠

get_active_producers⚠

get_blockchain_parameters_packed⚠

get_context_free_data⚠

get_permission_last_used⚠

get_resource_limits⚠

has_auth⚠

is_account⚠

is_privileged⚠

printdf⚠

printhex⚠

printi⚠

printi128⚠

printn⚠

printqf⚠

prints⚠

prints_l⚠

printsf⚠

printui⚠

printui128⚠

publication_time⚠

read_action_data⚠

read_transaction⚠

recover_key⚠

require_auth⚠

require_auth2⚠

require_recipient⚠

ripemd160⚠

send_context_free_inline⚠

send_deferred⚠

send_inline⚠

set_blockchain_parameters_packed⚠

set_privileged⚠

set_proposed_producers⚠

set_resource_limits⚠

sha1⚠

sha256⚠

sha512⚠

tapos_block_num⚠

tapos_block_prefix⚠

transaction_size⚠

Type Definitions

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

capi_name

int8_t

int16_t

int32_t

int64_t

int128_t

intmax_t

intptr_t

ptrdiff_t

size_t

ssize_t

uint8_t

uint16_t

uint32_t

uint64_t

uint128_t

uintmax_t

uintptr_t